pyproject.toml
[project]
name = "fastapi-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.8",
    "tortoise-orm>=0.24.0",
    "uvicorn>=0.34.0",
]
